Move NM to the notification package and avoid using VSCode interfaces
